requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.0
importlib-metadata
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.3
mysql-connector-python==9.1.0
Werkzeug==3.1.3
zipp